+++ benchmarks/index.js
@@ -14,9 +14,9 @@ const debug = require('debug')('benchmark');
 
 const benchmarks = new Benchmark({
 	// TODO: Update this once the module is published to as @prometheus/client.
-	'@prometheus/client@latest': 'prom-client@latest',
-	'@prometheus/client@trunk': 'git@github.com:prometheus/client_js',
-	'@prometheus/client@current': { location: process.cwd() },
+	released: 'prom-client@latest',
+	trunk: 'git@github.com:prometheus/client_js',
+	current: { location: process.cwd() },
 });
 
 benchmarks.suite('counter', require('./counter'));